test: parallelize long-running test
